refactoring location of files to make navigating the proposed infectious interface/ disease superclass easier
