+++ Protecs/src/main/java/uk/ac/ucl/protecs/behaviours/diseaseProgression/CoronavirusBehaviourFramework.java
@@ -1,4 +1,4 @@
-package uk.ac.ucl.protecs.behaviours;
+package uk.ac.ucl.protecs.behaviours.diseaseProgression;
