+++ Protecs/src/main/java/uk/ac/ucl/protecs/sim/WorldBankCovid19Sim.java
@@ -7,7 +7,7 @@ import java.util.Random;
 
 import uk.ac.ucl.protecs.behaviours.*;
 import uk.ac.ucl.protecs.objects.diseases.CoronavirusInfection;
-import uk.ac.ucl.protecs.behaviours.DummyBehaviourFramework;
+import uk.ac.ucl.protecs.behaviours.diseaseProgression.DummyBehaviourFramework;
 import uk.ac.ucl.protecs.objects.diseases.DummyInfection;
 import uk.ac.ucl.protecs.objects.diseases.Infection;
 import uk.ac.ucl.protecs.objects.hosts.Person;
@@ -16,11 +16,10 @@ import uk.ac.ucl.protecs.objects.hosts.Person.SEX;
 import uk.ac.ucl.protecs.objects.locations.Household;
 import uk.ac.ucl.protecs.objects.locations.Location;
 import uk.ac.ucl.protecs.objects.locations.Workplace;
-import uk.ac.ucl.protecs.behaviours.SpuriousSymptomBehaviourFramework;
-import uk.ac.ucl.protecs.behaviours.CoronavirusBehaviourFramework;
+import uk.ac.ucl.protecs.behaviours.diseaseProgression.SpuriousSymptomBehaviourFramework;
+import uk.ac.ucl.protecs.behaviours.diseaseProgression.CoronavirusBehaviourFramework;
 import sim.engine.SimState;
 import sim.engine.Steppable;
-import sim.util.Proxiable;
 
 public class WorldBankCovid19Sim extends SimState {
 	// Create a boolean for developing disease modularity
